+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.3.0
+
+- This is a major release deprecating activity flags in the trading pair universe.
+  The multichain trading pair data is too big to include inactivate trading pairs (800k+ total trading pairs).
+  Thus, the pair universe set only contains [active trading pairs](https://tradingstrategy.ai/docs/programming/tracking.html) after this release, 
+  making the trading pair universe less than 100k trading pairs again, making it more feasible to download the data.
+
 # 0.2.14
 
 - Make `ChainId` database more deterministic by loading all supported chains once and only once